refactor packing
