+++ functionary/train/packing/assert_packing_loss.py
@@ -9,7 +9,9 @@ import random
 import math
 import typer
 import sys
+from transformers import MistralForCausalLM
 from packed_dataset import PackedDataset
+from monkey_path_get_unpad_data import get_unpad_data
 
 
 random.seed(1)
@@ -65,7 +67,7 @@ def compute_loss_of_model(
 
 
 def compute_loss_for_model_class(
-    pretrained_path: str, model_class: Any, tokenizer: Any, ds: Any
+    pretrained_path: str, tokenizer: Any, ds: Any
 ) -> Tuple[float, int]:
     """Compute the loss with model initilized from model_class
         also return the number of tokens for computing the loss
@@ -78,12 +80,16 @@ def compute_loss_for_model_class(
     Returns:
         _type_: _description_
     """
-    model = model_class.from_pretrained(
+    model = transformers.AutoModelForCausalLM.from_pretrained(
         pretrained_path,
         torch_dtype=torch.bfloat16,
         device_map="auto",
         attn_implementation="flash_attention_2",
     )
+
+    # In the training, we set use_cache=False, use_cache=True only takes effect at inference
+    model.config.use_cache = False
+
     if hasattr(model, "router_aux_loss_coef"):
         print("set au_coef=0")
         model.router_aux_loss_coef = (
@@ -142,28 +148,6 @@ def main(
 
     model_config = transformers.AutoConfig.from_pretrained(pretrained_path)
     config_type = type(model_config).__name__.lower()
-    if "mistral" in config_type:
-        print("model: Mistral ")
-        from mistral_monkey_patch import MistralForCausalLM
-
-        mk_model_class = MistralForCausalLM
-    elif "llama" in config_type:
-        print("model: Llama ")
-        from llama_monkey_patch import LlamaForCausalLM
-
-        mk_model_class = LlamaForCausalLM
-    elif "mixtral" in config_type:
-        # Mixtral requires this ?
-        tokenizer.padding_side = "left"
-        print("model: Mixtral")
-        from mixtral_monkey_patch import MixtralForCausalLM
-
-        mk_model_class = MixtralForCausalLM
-    else:
-        print(
-            f"{config_type} is not supported, currently we only support: Mistral, Mixtral, Llama"
-        )
-        sys.exit(1)
 
     ds = load_dataset("tatsu-lab/alpaca")["train"]
     # extract 100 random data points from ds
@@ -171,8 +155,8 @@ def main(
     indices = [i for i in range(size)]
     random.shuffle(indices)
 
-    # We randomly select 150 data points for computing loss
-    ex_ds = ds.select(indices[:100])
+    # We randomly select 50 data points for computing loss
+    ex_ds = ds.select(indices[:50])
     print("number of data points: ", len(ex_ds))
     original_columns = ex_ds.column_names
 
@@ -199,16 +183,30 @@ def main(
     packed_ds = PackedDataset(ex_ds, tokenizer, pack_length)
     packed_ds.stat()
 
-    mk_avg_loss, mk_token_count = compute_loss_for_model_class(
-        pretrained_path, mk_model_class, tokenizer, packed_ds
-    )
-    print("monkey_patched loss: ", mk_avg_loss)
-    original_model_class = transformers.AutoModelForCausalLM
-
     original_avg_loss, original_token_count = compute_loss_for_model_class(
-        pretrained_path, original_model_class, tokenizer, ex_ds
+        pretrained_path, tokenizer, ex_ds
     )
     print("original_loss: ", original_avg_loss)
+
+    if "mistral" in config_type:
+        print("model: Mistral ")
+        transformers.models.mistral.modeling_mistral._get_unpad_data = get_unpad_data
+    elif "llama" in config_type:
+        print("model: Llama ")
+        transformers.models.llama.modeling_llama._get_unpad_data = get_unpad_data
+    elif "mixtral" in config_type:
+        print("model: Mixtral")
+        transformers.models.mixtral.modeling_mixtral._get_unpad_data = get_unpad_data
+    else:
+        print(
+            f"{config_type} is not supported, currently we only support: Mistral, Mixtral, Llama"
+        )
+        sys.exit(1)
+
+    # compute the loss on packed dataset with monkey-patched model
+    mk_avg_loss, mk_token_count = compute_loss_for_model_class(
+        pretrained_path, tokenizer, packed_ds
+    )
 
     # Make sure that number of tokens used for computing loss are the same in original dataset and packed dataset
     assert (
